test: aumentando level php stan para 5

+++ app/Services/Auth/AuthService.php
@@ -99,7 +99,7 @@ class AuthService
             'user' => [
                 'name' => $user->name,
                 'id' => $user->id,
-                'salutation' => CalendarTools::salutation($user->name, date('H')),
+                'salutation' => CalendarTools::salutation($user->name, (int)date('H')),
                 'salary' => $user->salary,
                 'email' => $user->email,
                 'marketPlannerValue' => $user->market_planner_value,
